+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Designate funds as location use restricted. Refs UIF-482.
 * Assign locations to restricted funds. Refs UIF-484.
 * Allow user to recalculate budgets based on conducted transactions when desired. Refs UIF-491.
+* Incorrect Display of Expense Class Percent. Refs UIF-509.
 
 ## [5.0.0](https://github.com/folio-org/ui-finance/tree/v5.0.0) (2023-10-12)
 [Full Changelog](https://github.com/folio-org/ui-finance/compare/v4.0.4...v5.0.0)