+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Change history for ui-finance
 
-## 5.1.0 (IN PROGRESS)
+## 6.0.0 (IN PROGRESS)
 
 * Add donor info to Fund. Refs UIF-473.
 * Designate funds as location use restricted. Refs UIF-482.
@@ -8,7 +8,7 @@
 * Allow user to recalculate budgets based on conducted transactions when desired. Refs UIF-491.
 * Incorrect Display of Expense Class Percent. Refs UIF-509.
 * Make location association required when funds are restricted by location. Refs UIF-492.
-* Add "Unrelease encumbrance" action to encumbrance transaction view. Refs UIF-503.
+* *BREAKING* Add "Unrelease encumbrance" action to encumbrance transaction view. Refs UIF-503.
 
 ## [5.0.0](https://github.com/folio-org/ui-finance/tree/v5.0.0) (2023-10-12)
 [Full Changelog](https://github.com/folio-org/ui-finance/compare/v4.0.4...v5.0.0)